# config/swagger-output.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Le Café",
    "version": "1.0.0",
    "contact": {}
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/menu": {
      "get": {
        "tags": [
          "Menu"
        ],
        "summary": "Get All Menu",
        "description": "Get All Menu",
        "operationId": "getAllMenu",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "5"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "passion"
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "non-coffee"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Menu"
        ],
        "summary": "Add Menu",
        "description": "Add Menu",
        "operationId": "addMenu",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "category": {
                    "type": "string",
                    "example": "coffee"
                  },
                  "description": {
                    "type": "string",
                    "example": "plerkudanil"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  },
                  "name": {
                    "type": "string",
                    "example": "Tuna Puff"
                  },
                  "price": {
                    "type": "string",
                    "example": "10000"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/menu/1": {
      "put": {
        "tags": [
          "Menu"
        ],
        "summary": "Update Menu",
        "description": "Update Menu",
        "operationId": "updateMenu",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "category": {
                    "type": "string",
                    "example": "non-coffee"
                  },
                  "description": {
                    "type": "string",
                    "example": "lorem ipsum dolar sit amet"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  },
                  "name": {
                    "type": "string",
                    "example": "Kopi Ojan"
                  },
                  "price": {
                    "type": "string",
                    "example": "20000"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "patch": {
        "tags": [
          "Menu"
        ],
        "summary": "Delete Menu",
        "description": "Delete Menu",
        "operationId": "deleteMenu",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "Delete Menu": {
                  "value": ""
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/order": {
      "get": {
        "tags": [
          "Order"
        ],
        "summary": "Get Order by user id",
        "description": "Get Order by user id",
        "operationId": "getOrderByUserId",
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Order"
        ],
        "summary": "Create Order",
        "description": "Create Order",
        "operationId": "createOrder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "menuJson": {
                    "type": "object",
                    "properties": {
                      "mn_desc": {
                        "type": "string",
                        "example": "Cafe Latte is a coffee-based drink prepared by diluting coffee with steamed milk, typically in a 3:1 or 4:1 ratio."
                      },
                      "mn_id": {
                        "type": "number",
                        "example": 1
                      },
                      "mn_image": {
                        "type": "string",
                        "example": "/images/caffelate.jpg"
                      },
                      "mn_name": {
                        "type": "string",
                        "example": "Cafe Latte"
                      },
                      "mn_price": {
                        "type": "number",
                        "example": 30000
                      }
                    }
                  },
                  "quantity": {
                    "type": "number",
                    "example": 100
                  },
                  "site": {
                    "type": "string",
                    "example": "Tangerang Selatan"
                  },
                  "totalPrice": {
                    "type": "number",
                    "example": 800000
                  },
                  "typeOrder": {
                    "type": "string",
                    "example": "Delivery"
                  },
                  "userId": {
                    "type": "number",
                    "example": 1
                  }
                }
              },
              "examples": {
                "Create Order": {
                  "value": {
                    "menuJson": {
                      "mn_desc": "Cafe Latte is a coffee-based drink prepared by diluting coffee with steamed milk, typically in a 3:1 or 4:1 ratio.",
                      "mn_id": 1,
                      "mn_image": "/images/caffelate.jpg",
                      "mn_name": "Cafe Latte",
                      "mn_price": 30000
                    },
                    "quantity": 100,
                    "site": "Tangerang Selatan",
                    "totalPrice": 800000,
                    "typeOrder": "Delivery",
                    "userId": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
...
}